+++ src/main/java/chess/domain/Board.java
@@ -3,7 +3,7 @@ package chess.domain;
 import chess.domain.piece.Piece;
 import chess.domain.piece.Team;
 import chess.domain.position.Position;
-import chess.domain.util.PointCalculator;
+import chess.util.PointCalculator;
 import java.util.Map;
 
 public class Board {
