+++ src/test/java/chess/domain/ChessGameTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import chess.domain.piece.Pawn;
 import chess.domain.piece.Team;
 import chess.domain.position.Position;
-import chess.domain.util.BoardInitializer;
+import chess.util.BoardInitializer;
 import java.util.EnumMap;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
